services/ai-service/requirements.txt
flask==3.1.*
gunicorn==23.*
python-dotenv==1.*
openai==1.*
PyPDF2==3.*
pdfplumber==0.11.*
pydantic==2.*
